requirements.txt
pillow
streamlit
protobuf
pydantic
google-genai
python-dotenv
rembg
onnxruntime
requests
fastapi